Add files via upload
Smart E-Com is a modern, responsive e-commerce landing page designed to give users a smooth and engaging online shopping experience. The homepage features a clean layout with a vibrant blue header, a dynamic hero section, and floating shopping icons that add a lively touch. It introduces users to the brand with a bold headline, a short tagline, and an inviting “Get Started” button that directs them to the main shopping page. The dropdown menu provides easy access to essential links like login options.

Built using HTML, CSS, and JavaScript, the design ensures compatibility across devices, offering a user-friendly and visually appealing interface that reflects the efficiency and innovation of online shopping.