Add files via upload
The Smart E-Com website is a modern and responsive e-commerce homepage designed using HTML, CSS, and JavaScript. It features a stylish layout with a blue-themed header, interactive dropdown menu, and an attractive hero section that highlights the brand’s message — offering users a fast, easy, and enjoyable online shopping experience. The design includes floating shopping icons and a “Get Started” button that leads users to the main shopping page. The site is user-friendly, visually appealing, and adaptable to both desktop and mobile screens.